finish

+++ app/Http/Controllers/Admin/Type_product_Controller.php
@@ -20,7 +20,7 @@ class Type_product_Controller extends Controller
     }
 
     public function create(Request $request){
-        
+
         $validatedData = $request->validate([
             'name' => 'required|unique:type_products|max:255'
         ],
@@ -35,7 +35,7 @@ class Type_product_Controller extends Controller
         $create_type->name = $request->name;
         $create_type->id_admin = Auth::user()->id;
         $create_type->save();
-        return redirect('/Admin/type_product/index');
+        return redirect('/Admin/type_product/index')->with('success','บันทึกข้อมูลเรียบร้อย');
     }
 
 
@@ -48,11 +48,10 @@ class Type_product_Controller extends Controller
     public function update(Request $request,$id_type){
 
         $validatedData = $request->validate([
-            'name' => 'required|unique:type_products|max:255'
+            'name' => 'required|max:255'
         ],
         [
             'name.required' => 'กรุณกรอกประเภทสินค้า',
-            'name.unique' => 'มีประเภทสินค้านี้แล้ว',
             'name.max:255' => 'กรอกเกิน 255 ตัวอักษร'
         ]
         );
@@ -60,12 +59,16 @@ class Type_product_Controller extends Controller
         $update_type = Type_product::find($id_type);
         $update_type->name = $request->name;
         $update_type->save();
-        return redirect('/Admin/type_product/index');
+        return redirect('/Admin/type_product/index')->with('success','บันทึกข้อมูลเรียบร้อย');
     }
 
     //delete
     public function delete($id_type){
+        $type = Type_product::find($id_type);
+        if($type->product->count() > 0){
+            return redirect()->back()->with('error','ไม่สามารถลบประเภทสินค้าได้เนื่องจากมีสินค้าอยู่');
+        }
         Type_product::destroy($id_type);
-        return redirect('/Admin/type_product/index');
+        return redirect('/Admin/type_product/index')->with('delete','ลบข้อมูลเรียบร้อย');
     }
 }
